+++ sophonsoar-20220728/alibabacloud_sophonsoar20220728/client.py
@@ -4285,7 +4285,7 @@ class Client(OpenApiClient):
         """
         @summary Submits and runs a Python3 script. You can call this operation only for data processing.
         
-        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=openapi-amp.newDocPublishment.0.0.4c41281fWhbdPa#/commodity/vm_intl).
         
         @param request: RunPython3ScriptRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4328,7 +4328,7 @@ class Client(OpenApiClient):
         """
         @summary Submits and runs a Python3 script. You can call this operation only for data processing.
         
-        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=openapi-amp.newDocPublishment.0.0.4c41281fWhbdPa#/commodity/vm_intl).
         
         @param request: RunPython3ScriptRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4370,7 +4370,7 @@ class Client(OpenApiClient):
         """
         @summary Submits and runs a Python3 script. You can call this operation only for data processing.
         
-        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=openapi-amp.newDocPublishment.0.0.4c41281fWhbdPa#/commodity/vm_intl).
         
         @param request: RunPython3ScriptRequest
         @return: RunPython3ScriptResponse
@@ -4385,7 +4385,7 @@ class Client(OpenApiClient):
         """
         @summary Submits and runs a Python3 script. You can call this operation only for data processing.
         
-        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing method and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=openapi-amp.newDocPublishment.0.0.4c41281fWhbdPa#/commodity/vm_intl).
         
         @param request: RunPython3ScriptRequest
         @return: RunPython3ScriptResponse
@@ -4401,7 +4401,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers an enabled custom playbook or a predefined playbook.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerPlaybookRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4440,7 +4440,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers an enabled custom playbook or a predefined playbook.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerPlaybookRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4478,7 +4478,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers an enabled custom playbook or a predefined playbook.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerPlaybookRequest
         @return: TriggerPlaybookResponse
@@ -4493,7 +4493,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers an enabled custom playbook or a predefined playbook.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerPlaybookRequest
         @return: TriggerPlaybookResponse
@@ -4613,7 +4613,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers a playbook or a command.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerSophonPlaybookRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4658,7 +4658,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers a playbook or a command.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerSophonPlaybookRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4702,7 +4702,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers a playbook or a command.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerSophonPlaybookRequest
         @return: TriggerSophonPlaybookResponse
@@ -4717,7 +4717,7 @@ class Client(OpenApiClient):
         """
         @summary Triggers a playbook or a command.
         
-        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.aliyun.com/price/product#/sas/detail/sas).
+        @description Before you call this operation, make sure that you understand the billing methods and pricing of Security Orchestration Automation Response (SOAR). For more information, see [Pricing](https://www.alibabacloud.com/en/pricing-calculator?_p_lc=1&spm=a2796.7960336.3034855210.1.7adab91arMeIx2#/commodity/vm_intl).
         
         @param request: TriggerSophonPlaybookRequest
         @return: TriggerSophonPlaybookResponse