+++ sophonsoar-20220728/ChangeLog.md
@@ -1,3 +1,14 @@
+2024-11-02 Version: 2.0.1
+- Update API CreatePlaybook: add param TaskflowType.
+- Update API DescribeDistinctReleases: update response param.
+- Update API DescribePlaybooks: add param Order.
+- Update API DescribePlaybooks: add param ParamTypes.
+- Update API DescribePlaybooks: add param Sort.
+- Update API DescribePlaybooks: update response param.
+- Update API DescribeProcessTasks: add param EntityUuid.
+- Update API DescribeProcessTasks: update response param.
+
+
 2024-06-28 Version: 2.0.0
 - Update API CreatePlaybook: add param TaskflowType.
 - Update API DescribeDistinctReleases: update response param.